:art: Fixes type errors with tree
